[IMP] Refactoring: splitted files for each new class
